Finish moving things around
