+++ relay-event-normalization/src/normalize/span/mod.rs
@@ -1,5 +1,6 @@
 //! Span normalization logic.
 
+pub mod ai;
 pub mod description;
 pub mod exclusive_time;
 pub mod tag_extraction;
