+++ relay-server/src/metrics_extraction/event.rs
@@ -69,8 +69,9 @@ fn extract_span_metrics_for_event(
     max_tag_value_size: usize,
     output: &mut Vec<Bucket>,
 ) {
+    println!("Extracting span metrics");
     relay_statsd::metric!(timer(RelayTimers::EventProcessingSpanMetricsExtraction), {
-        if let Some(transaction_span) = extract_transaction_span(event, max_tag_value_size, None) {
+        if let Some(transaction_span) = extract_transaction_span(event, max_tag_value_size) {
             output.extend(generic::extract_metrics(&transaction_span, config));
         }
