+++ relay-server/src/services/processor/span.rs
@@ -1,6 +1,6 @@
 //! Processor code related to standalone spans.
 
-use relay_dynamic_config::{Feature, ModelCosts};
+use relay_dynamic_config::Feature;
 use relay_event_normalization::span::tag_extraction;
 use relay_event_schema::protocol::{Event, Span};
 use relay_protocol::Annotated;
@@ -10,7 +10,6 @@ use crate::{services::processor::ProcessEnvelopeState, utils::ItemAction};
 
 #[cfg(feature = "processing")]
 mod processing;
-use crate::services::processor::ai::extract_ai_measurements;
 #[cfg(feature = "processing")]
 pub use processing::*;
 
@@ -31,16 +30,9 @@ pub fn filter(state: &mut ProcessEnvelopeState<SpanGroup>) {
 /// Creates a span from the transaction and applies tag extraction on it.
 ///
 /// Returns `None` when [`tag_extraction::extract_span_tags`] clears the span, which it shouldn't.
-pub fn extract_transaction_span(
-    event: &Event,
-    max_tag_value_size: usize,
-    ai_model_costs: Option<ModelCosts>,
-) -> Option<Span> {
-    let mut span = Span::from(event);
-    if let Some(model_costs) = ai_model_costs {
-        extract_ai_measurements(&mut span, &model_costs)
-    }
-    let mut spans = [span.into()];
+pub fn extract_transaction_span(event: &Event, max_tag_value_size: usize) -> Option<Span> {
+    let mut spans = [Span::from(event).into()];
+
     tag_extraction::extract_span_tags(event, &mut spans, max_tag_value_size);
 
     spans.into_iter().next().and_then(Annotated::into_value)
